Separation of concerns: Users and posts
