+++ apps/multirotor/lib/posts/posts.ex
@@ -18,24 +18,14 @@ defmodule Multirotor.Posts do
   @required_fields ~w(title,body,date,userid,type)a
   @optional_fields ~w(votes,views)a
 
-
   def changeset(post, params \\ %{}) do
     post
     |> cast(params, [:title, :body, :votes, :date, :views, :userid, :type])
         |> validate_required([:title, :body, :date, :userid, :type])
-
     #|> validate_required(@required_fields)
-
   end
 #|> validate_change(:date, &must_be_future/2)
  #For example purposes
-  defp must_be_future(_, value) do
-    Ecto.DateTime.compare(value, Ecto.DateTime.utc)
-    |> get_error
-  end
-
-  defp get_error(comparison) when comparison == :lt, do: [date: "Cannot be in the past"]
-  defp get_error(_), do: []
-
+  
 
 end
