+++ apps/multirotor_web/web/controllers/sessions_controller.ex
@@ -10,7 +10,7 @@ defmodule MultirotorWeb.SessionsController do
 
   def identity_callback(%{assigns: %{ueberauth_auth: auth}} = conn, _) do
     conn
-    |> authenticated(MultirotorUsers.Users.authenticate auth)
+    |> authenticated(MultirotorUsers.UserAuthentication.authenticate auth)
   end
 
   def delete(conn, _params) do
